+++ R/growthLogistic.R
@@ -9,7 +9,7 @@
 #' @param msg A logical that indicates whether a message about the model and parameter definitions should be output (\code{=TRUE}) or not (\code{=FALSE}; default).
 #' @param \dots Not implemented.
 #' 
-#' @return \code{logisticFuns} returns a function that can be used to predict fish length or weight given a vector of ages and values for the function parameters.  The result should be saved to an object that can then be used as a function name.  When the resulting function is used, the parameters are ordered as shown when the definitions of the parameters are printed after the function is called (if \code{msg=TRUE}).  #'If \code{simple=FALSE}, then the values for all parameters may be included as a vector in the first parameter argument.  If \code{simple=TRUE}, then all parameters must be declared individually.  The resulting function is somewhat easier to read when \code{simple=TRUE}.
+#' @return \code{logisticFuns} returns a function that can be used to predict fish length or weight given a vector of ages and values for the function parameters.  The result should be saved to an object that can then be used as a function name.  When the resulting function is used, the parameters are ordered as shown when the definitions of the parameters are printed after the function is called (if \code{msg=TRUE}).  If \code{simple=FALSE}, then the values for all parameters may be included as a vector in the first parameter argument.  If \code{simple=TRUE}, then all parameters must be declared individually.  The resulting function is somewhat easier to read when \code{simple=TRUE}.
 #' 
 #' \code{logisticModels} returns a graphic that uses \code{\link{plotmath}} to show the model formulae in a pretty format.
 #'
@@ -19,12 +19,14 @@
 #'
 #' @section IFAR Chapter: None specifically, but \href{https://fishr.wordpress.com/books/ifar/}{9-Individual Growth} is related.
 #'
-#' @seealso See \code{\link{vbFuns}}, \code{\link{gompFuns}}, and \code{\link{schnute}} for similar functionality for other models.
+#' @seealso See \code{\link{vbFuns}}, \code{\link{gompFuns}}, \code{\link{logisticFuns}}, and \code{\link{schnute}} for similar functionality for other models.
 #'
 #' @references 
-#' Campana, S.E. and C.M. Jones.  1992.  Analysis of otolith microstructure data.  Pages 73-100 In D.K. Stevenson and S.E. Campana, editors.  Otolith microstructure examination and analysis.  Canadian Special Publication of Fisheries and Aquatic Sciences 117.
+#' Campana, S.E. and C.M. Jones.  1992.  \href{http://www.dfo-mpo.gc.ca/Library/141734.pdf}{Analysis of otolith microstructure data}.  Pages 73-100 In D.K. Stevenson and S.E. Campana, editors.  Otolith microstructure examination and analysis.  Canadian Special Publication of Fisheries and Aquatic Sciences 117.
 #' 
-#' Winsor, C.P.  1932.  \href{http://www.ncbi.nlm.nih.gov/pmc/articles/PMC1076153/pdf/pnas01729-0009.pdf}{The Gompertz curve as a growth curve}.  Proceedings of the National Academy of Sciences.  18:1-8.
+#' Haddon, M., C. Mundy, and D. Tarbath. 2008.  \href{http://aquaticcommons.org/8857/1/haddon_Fish_Bull_2008.pdf}{Using an inverse-logistic model to describe growth increments of blacklip abalone (\emph{Haliotis rubra}) in Tasmania}.  Fishery Bulletin 106:58-71. 
+#' 
+#' Karkach, A. S.  2006.  \href{http://www.demographic-research.org/volumes/vol15/12/15-12.pdf}{Trajectories and models of individual growth}.  Demographic Research 15:347-400.
 #' 
 #' @keywords manip
 #'
@@ -42,6 +44,11 @@
 #' plot(log2(ages,Linf=800,gninf=0.5,a=10)~ages,type="b",pch=19)
 #'
 #' ( log2c <- logisticFuns("CJ2",simple=TRUE) ) # compare to log2
+#' 
+#' ( log3 <- logisticFuns("Karkach") )       # Karkach parametrization
+#' plot(log3(ages,L0=10,Linf=800,gninf=0.5)~ages,type="b",pch=19)
+#' 
+#' ( log4 <- logisticFuns("HaddonI") )       # Haddon's inverse logistic
 #'
 #' #######################################################################################
 #' ## Examples of fitting Logistic models
@@ -73,8 +80,8 @@
 #' summary(fit2,correlation=TRUE)
 #' curve(log2(x,Linf=coef(fit2)),from=0,to=15,col="blue",lwd=5,add=TRUE)
 #'
-#' # Fit Richards-like parameterization (using simple=TRUE model)
-#' log3 <- logisticFuns("Richards",simple=TRUE)
+#' # Fit Karkach parameterization (using simple=TRUE model)
+#' log3 <- logisticFuns("Karkach",simple=TRUE)
 #' fit3 <- nls(len~log3(age,Linf,L0,gninf),data=df,start=list(Linf=450,L0=30,gninf=0.45))
 #' summary(fit3,correlation=TRUE)
 #' curve(log3(x,Linf=coef(fit3)[1],L0=coef(fit3)[2],gninf=coef(fit3)[3]),
@@ -84,7 +91,7 @@ NULL
 
 #' @rdname growthLogistic
 #' @export
-logisticFuns <- function(type=c("CJ1","CJ2","Richards"),simple=FALSE,msg=FALSE) {
+logisticFuns <- function(type=c("CJ1","CJ2","Karkach","HaddonI"),simple=FALSE,msg=FALSE) {
   CJ1 <- function(t,Linf,gninf=NULL,ti=NULL) {
     if (length(Linf)==3) { gninf <- Linf[[2]]
                            ti <- Linf[[3]]
@@ -103,14 +110,23 @@ logisticFuns <- function(type=c("CJ1","CJ2","Richards"),simple=FALSE,msg=FALSE) 
   SCJ2<- function(t,Linf,gninf,a) {
     Linf/(1+a*exp(-gninf*t))
   }
-  Richards <- function(t,Linf,L0=NULL,gninf=NULL) {
+  Karkach <- function(t,Linf,L0=NULL,gninf=NULL) {
     if (length(Linf)==3) { L0 <- Linf[[2]]
                            gninf <- Linf[[3]]
                            Linf <- Linf[[1]] }
     L0*Linf/(L0+(Linf-L0)*exp(-gninf*t))
   }
-  SRichards <- function(t,Linf,L0,gninf) {
+  SKarkach <- function(t,Linf,L0,gninf) {
     L0*Linf/(L0+(Linf-L0)*exp(-gninf*t))
+  }
+  HaddonI <- function(Lm,dLmax,L50=NULL,L95=NULL) {
+    if (length(dLmax)==3) { L50=dLmax[2]
+                            L95=dLmax[3]
+                            dLmax=dLmax[1] }
+    dLmax/(1+exp(log(19)*((Lm-L50)/(L95-L50))))
+  }
+  SHaddonI <- function(Lm,dLmax,L50,L95) {
+    dLmax/(1+exp(log(19)*((Lm-L50)/(L95-L50))))
   }
   ## Main function
   type <- match.arg(type)
@@ -131,12 +147,21 @@ logisticFuns <- function(type=c("CJ1","CJ2","Richards"),simple=FALSE,msg=FALSE) 
                 "          gi = instantaneous growth rate at the inflection point\n",
                 "           a = a dimensionless parameter related to growth\n\n")
       },
-      Richards= {
-        message("You have chosen the 'Richards'",comcat,
+      Karkach= {
+        message("You have chosen the 'Karkach'",comcat,
                 "  E[L|t] = L0*Linf/(L0+(Linf-L0)*exp(-gninf*t))\n\n",
                 "  where Linf = asymptotic mean length\n",
                 "          L0 = mean length at time/age 0\n",
                 "          gi = instantaneous growth rate at the inflection point\n\n")
+      },
+      HaddonI= {
+        message("You have chosen the 'Haddon Inverse'",comcat,
+                "  E[Lr-Lm|dt] = dLmax/(1+exp(log(19)*((Lm-L50)/(L95-L50))))\n\n",
+                "  where dLmax = maximum growth increment during the study\n",
+                "          L50 = length at marking to produces a growth increment of 0.5*dLmax",
+                "          L95 = length at marking to produces a growth increment of 0.95*dLmax\n\n",
+                "  and the data are Lr = length at time of recapture\n",
+                "                   Lm = length at time of marking\n")
       }
     )
   }
@@ -151,7 +176,7 @@ logisticModels <- function(...) {
   plot(1,type="n",ylim=c(0,3),xlim=c(0,1),xaxt="n",yaxt="n",xlab="",ylab="",bty="n",main="FSA Logistic Growth Parameterizations",...)
   iGrowthModels("CJ1", 0.1,2.5)
   iGrowthModels("CJ2", 0.1,1.5)
-  iGrowthModels("Richards", 0.1,0.5)
+  iGrowthModels("Karkach", 0.1,0.5)
   par(op)
 }